Move all files to `src`, removing `function` and `cli` folders
