+++ src/checkExtendedConfig.ts
