+++ src/computeFileContents.ts
