+++ src/extendArraysOfObjects.ts
