+++ src/extendDotIgnore.ts
