--- src/functions/types.ts
@@ -1,1 +1,0 @@
-export type ExtendJson = (destination: any, source: any) => any;
