--- src/typings/fs-monkey.d.ts
@@ -1,4 +1,0 @@
-declare module 'fs-monkey' {
-  export const patchFs: (a: any, b: any) => any;
-  export const patchFs: (a: any) => any;
-}
